requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
bcrypt==4.0.1
bleach==6.0.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
click-log==0.4.0
colorama==0.4.6 ;
cryptography==39.0.2
deprecation==2.1.0
docutils==0.19 ; python_version >= '3.7'
dotty-dict==1.3.1 ; python_version >= '3.5' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.94.0
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.31 ; python_version >= '3.7'
gotrue==1.0.0 ; python_version >= '3.8' and python_version < '4.0'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version >= '3.7'
invoke==1.7.3
jaraco.classes==3.2.3 ; python_version >= '3.7'
keyring==23.13.1 ; python_version >= '3.7'
markdown==3.4.1
more-itertools==9.1.0 ; python_version >= '3.7'
packaging==23.0 ; python_version >= '3.7'
passlib==1.7.4
pillow==9.4.0
pkginfo==1.9.6 ; python_version >= '3.6'
postgrest==0.10.6 ; python_version >= '3.8' and python_version < '4.0'
pyasn1==0.4.8
pycparser==2.21
pydantic[dotenv]==1.10.6
pygments==2.14.0 ; python_version >= '3.6'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-gitlab==3.13.0 ; python_full_version >= '3.7.0'
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-semantic-release==7.33.2
pywin32-ctypes==0.2.0 ; sys_platform == 'win32'
readme-renderer==37.3 ; python_version >= '3.7'
realtime==1.0.0 ; python_version >= '3.8' and python_version < '4.0'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-toolbelt==0.10.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3986[idna2008]==1.5.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
semver==2.13.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.26.0.post1 ; python_version >= '3.7'
storage3==0.5.2 ; python_version >= '3.8' and python_version < '4.0'
strenum==0.4.9
supabase==1.0.2
supafunc==0.2.2 ; python_version >= '3.7' and python_version < '4.0'
tomlkit==0.11.6 ; python_version >= '3.6'
tqdm==4.65.0 ; python_version >= '3.7'
twine==3.8.0 ; python_version >= '3.6'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.21.0
webencodings==0.5.1
websockets==10.4 ; python_version >= '3.7'
wheel==0.38.4 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
